staging: android: lmk: Fix selectively count free CMA pages
Fix lowmemorykiller.c according to commit
3f1e551a0ea35715dfcbda2fc15c466e3cca6d42 in CAF KK_3.5.

In certain memory configurations there can be a large number of CMA
pages which are not suitable to satisfy certain memory requests.

This large number of unsuitable pages can cause the lowmemorykiller to
not kill any tasks because the lowmemorykiller counts all free pages. In
order to ensure the lowmemorykiller properly evaluates the free memory
only count the free CMA pages if they are suitable for satisfying the
memory request.

Change-Id: I7f06d53e2d8cfe7439e5561fe6e5209ce73b1c90 CRs-fixed: 437016
Signed-off-by: Liam Mark <lmark@codeaurora.org>